requirements.txt
langgraph
langchain-groq
langchain-community
llama-parse
pydantic
python-dotenv
instructor
fastapi
uvicorn